Exercises 4.20 - 4.21 completed

+++ controllers/blogs.js
@@ -8,22 +8,14 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs)
 })
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.startsWith('Bearer ')) {
-        return authorization.replace('Bearer ', '')
-    }
-    return null
-}
-
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
 
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!decodedToken.id) {
         return response.status(401).json({ error: 'token invalid'})
     }
-    
+
     const user = await User.findById(decodedToken.id)
 
     if (!body.title || !body.url) {
@@ -31,8 +23,6 @@ blogsRouter.post('/', async (request, response) => {
     }
 
     const likes = body.likes !== undefined ? body.likes : 0
-
-    //const user = await User.findById(body.userId)
 
     const blog = new Blog({
         title: body.title,
@@ -50,8 +40,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid'})
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        return response.status(401).json({ error: 'user unauthorized'})
+    }
+
+    
 })
 
 blogsRouter.put('/:id', async (request,response) => {
